// frontend/src/app/core/components/filters/filter-key.enum.ts
export enum FilterKey {
  SEARCH = 'SEARCH',

  //ITEMS
  ITEM_TYPE = 'ITEM_TYPE',
  WEAPON_DIFFICULTY = 'WEAPON_DIFFICULTY',
  WEAPON_CATEGORY = 'WEAPON_CATEGORY',
  ARMOR_CATEGORY = 'ARMOR_CATEGORY',
  TOOL_CATEGORY = 'TOOL_CATEGORY',
  EXPENDABLE = 'EXPENDABLE',
  EQUIPPABLE = 'EQUIPPABLE',
  CONTAINER = 'CONTAINER',
  MAGICAL_TYPE = 'MAGICAL_TYPE',
  RARITY = 'RARITY',
  ATTUNEMENT = 'ATTUNEMENT',
  CURSED = 'CURSED',

  //FEATURES
  FEATURE_CATEGORY = 'FEATURE_CATEGORY',
  FEATURE_CLASS = 'FEATURE_CLASS',
  FEATURE_RACE = 'FEATURE_RACE',
  FEATURE_BACKGROUND = 'FEATURE_BACKGROUND',
  LEVEL = 'LEVEL',
  RANGE = 'RANGE',
  AREA_OF_EFFECT = 'AREA_OF_EFFECT',
  FEATURE_AREA_OF_EFFECT = 'FEATURE_AREA_OF_EFFECT',
  ACTIVE = 'ACTIVE',
  PASSIVE = 'PASSIVE',

  //SPELLS
  PREPARED = 'PREPARED',
  CONCENTRATING = 'CONCENTRATING',
  SPELL_CLASS = 'SPELL_CLASS',
  SCHOOL = 'SCHOOL',
  RITUAL = 'RITUAL',
  SPELL_AREA_OF_EFFECT = 'SPELL_AREA_OF_EFFECT',
  VERBAL = 'VERBAL',
  SOMATIC = 'SOMATIC',
  MATERIAL = 'MATERIAL',
  CONCENTRATION = 'CONCENTRATION',
  INSTANTANEOUS = 'INSTANTANEOUS',
  TAGS = 'TAGS',

  //MONSTERS
  MONSTER_TYPE = 'MONSTER_TYPE',
  CHALLENGE_RATING = 'CHALLENGE_RATING',
  ALIGNMENT = 'ALIGNMENT',
  SPELLCASTER = 'SPELLCASTER',
  LEGENDARY = 'LEGENDARY',
  FLYING = 'FLYING',
  SWIMMING = 'SWIMMING'
}